fluoride: fix asan report error
btif/src/btif_dm.cc:1440:44: runtime error: shift exponent 32 is too large for 32-bit type 'int'
btif/src/btif_storage.cc:665:48: runtime error: shift exponent 32 is too large for 32-bit type 'int'
btcore/src/property.cc:188:46: runtime error: load of misaligned address 0xf3e0516f for type 'const bt_device_type_t', which requires 4 byte alignment

+++ btif/src/btif_storage.cc
@@ -662,7 +662,7 @@ bt_status_t btif_storage_get_adapter_property(bt_property_t* property) {
     for (i = 0; i < BTA_MAX_SERVICE_ID; i++) {
       /* This should eventually become a function when more services are enabled
        */
-      if (service_mask & (tBTA_SERVICE_MASK)(1 << i)) {
+      if (service_mask & (tBTA_SERVICE_MASK)((uint64_t)1 << i)) {
         switch (i) {
           case BTA_HFP_SERVICE_ID: {
             *(p_uuid + num_uuids) =
